Move files so that repo works as a Python library
